[fix] fix mutation to update variables with float values (#1876)
Quedó pendiente modificar la mutación del update para las variables de
la petición.

+++ server/src/db/repositories/PetitionRepository.ts
@@ -7640,7 +7640,7 @@ export class PetitionRepository extends BaseRepository {
         variables = (
         select jsonb_agg(
           case
-            when element->>'name' = ? then jsonb_build_object('name', ?::text, 'default_value', ?::int)
+            when element->>'name' = ? then jsonb_build_object('name', ?::text, 'default_value', ?::float)
             else element
           end
         )
